Annexes/B-Format_pivot_XML_TEI_Lectaurep/generator_Lectaurep2TEI/requirements.txt
lxml
bs4
pyfiglet
pillow
tqdm
termcolor